+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.2.0
+- Renamed package to Bishop
+- Piece drops & hands (Crazyhouse support)
+
 ### 0.1.1
 - Zobrist hashing & repetition draws